requirements.txt
# Core dependencies for Voice-Based Agentic RAG System
pydantic==2.12.5
PyYAML==6.0.3
pytest==9.0.2
pytest-cov==7.1.0
gradio==6.10.0
openai==2.31.0
tiktoken==0.12.0
typer==0.23.1
rich==14.3.3

# Voice input layer (Phase 1.2)
faster-whisper==1.2.1
silero-vad==6.2.1
sounddevice==0.5.5
PyAudio==0.2.14

# Text-to-speech (Phase 1.9)
pyttsx3==2.99
pygame==2.6.1
gTTS==2.5.4
kokoro==0.9.4

# Document parsing (Phase 1.5)
chardet==7.4.0.post2
mistune==3.2.0
python-docx==1.2.0
PyMuPDF==1.27.2.2
pandas==3.0.2
pdfplumber==0.11.9
beautifulsoup4==4.14.3

# Additional parsing formats (Phase 2.1)
openpyxl==3.1.5
xlrd==2.0.2
lxml==6.0.2
trafilatura==2.0.0
EbookLib==0.20
python-pptx==1.0.2
odfpy==1.4.1
chromadb==1.5.7
qdrant-client==1.17.1
rank-bm25==0.2.2
sentence-transformers==5.4.0
